requirements.txt
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.8.0
python-dotenv==1.0.1

# LLM & Embeddings
groq==0.9.0
sentence-transformers==3.0.1
torch==2.3.0

# Database
supabase==2.5.0
psycopg2-binary==2.9.9
pgvector==0.3.2

# Web Search
google-search-results==2.4.2

# Data Processing
beautifulsoup4==4.12.3
langchain-text-splitters==0.2.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Email Service
resend==0.8.0

# Utilities
requests==2.32.3
aiohttp==3.9.5